added new version of b9 cuda

+++ projects/resources/cuda/multi_gpu/b9.cuh
@@ -43,10 +43,10 @@ class Benchmark9M : public Benchmark {
    private:
     int S;
 
-    float **A, **t1, **t2;
-    float *x, *y, *r, *p, *b;
-    float t1_tot = 0;
-    float t2_tot = 0;
+    double **A, **t1, **t2;
+    double *x, *y, *r, *p, *b;
+    double t1_tot = 0;
+    double t2_tot = 0;
 
     cudaStream_t s1, s2;
     cudaStream_t *s;
